+++ lib/src/widgets/calendar_grid.dart
@@ -93,7 +93,9 @@ class CalendarGrid extends StatelessWidget {
                   final day = c.days[i];
                   final holiday = getDayHoliday(day);
                   final child = Container(
-                    padding: EdgeInsets.all(3),
+                    width: 22,
+                    height: 22,
+                    alignment: Alignment.center,
                     decoration: BoxDecoration(
                       shape: BoxShape.circle,
                       color:
@@ -148,9 +150,9 @@ class CalendarGrid extends StatelessWidget {
                           height: double.infinity,
                           width: 1,
                           color:
-                              (i < c.days.length - 1)
-                                  ? Colors.grey
-                                  : Colors.transparent,
+                              c.dateToHighlight(day)
+                                  ? context.watch<GanttTheme>().defaultCellColor
+                                  : Colors.grey,
                         ),
                       ],
                     ),